gcc14: param-type-aware carrier casts on the cps->cps call path
Second path of the carrier-to-typed-param front. A cps->cps call threads args to
a `__cps` callee whose value params are NOT uniformly int64 -- some are `void *`,
some concrete `tur_adt_X *` (emit_params spells each via binder_ctype_full). The
old atoms_csv_call only cast TY_FN args, so a pointer arg into an int64 param, or
an int64-carried arg into a void*/concrete-pointer param, tripped
-Wint-conversion (a hard error under GCC >= 14). A blanket int64 cast is wrong
(it breaks void* params).

Added atoms_csv_call_typed: for each arg it computes the callee's DECLARED param
C type (cps_call_param_ctype, mirroring emit_params exactly) and casts through
intptr_t only on a genuine mismatch -- skipping the provably-valid pointer->void*
case and leaving matching args bare (minimal churn). By-value aggregate args
(whose C type is neither int64_t nor a pointer) are left to the existing spill
bridge, since `(int64_t)(intptr_t)(agg)` is invalid; TY_FN args stay scalar.
Used at both cps->cps sites (main tailcall + heap-join).

Fixes 3 fixtures (list-basic, hamt-lisp-map-filter, hamt-lisp-merge-with);
278-line cast-only snapshot churn; full suite green (2202 passed, 0 failed).
Remaining paths (cps->direct, existential ctor, extern-c, by-value aggregate
element, stdlib inline-C) tracked in the report.

+++ tests/fixtures/mutex-snapshot/expected.c
@@ -4965,7 +4965,7 @@ static intptr_t map_hyeq_hyloop_j3(intptr_t env, intptr_t __t1__slot, DK *__kont
         /* panic-return-signal: ret ctype unknown; no propagation here */
         __t2 = __ps_68;
         if (__t2) {
-            return map_hyeq_hyloop__cps(iter, m2_hamt, keyeq, val_cmp, __kont); /* cps->cps */
+            return map_hyeq_hyloop__cps(iter, m2_hamt, keyeq, (int64_t)(intptr_t)val_cmp, __kont); /* cps->cps */
         } else {
             return dk_run(__kont, (intptr_t)(false));
         }
@@ -5361,7 +5361,7 @@ static int64_t list_hyeq_qu__cps(int64_t l1, int64_t l2, int64_t cmp_fn, DK *__k
                 __auto_type __ps_91 = (list_hytail(l2));
                 /* panic-return-signal: ret ctype unknown; no propagation here */
                 __t5 = __ps_91;
-                return list_hyeq_qu__cps(__t4, __t5, cmp_fn, __kont); /* cps->cps */
+                return list_hyeq_qu__cps(__t4, __t5, (int64_t)(intptr_t)cmp_fn, __kont); /* cps->cps */
             } else {
                 return dk_run(__kont, (intptr_t)(false));
             }
